meeting-intelligence/requirements.txt
google-generativeai>=0.3.0
langgraph>=0.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.8.0
httpx>=0.27.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
pydantic>=2.0.0